refactor(vm): split local.go and ansible.go, consolidate splitHosts
Two oversized files in the vm module are split into focused units, all
inside the same package — no caller changes:

local.go (490 LOC) →
  - local.go     — BakeLocal (kept)
  - profile.go   — BakeLocalByProfile + ProfileConfig type

ansible.go (511 LOC) →
  - ansible.go         — ExecuteAnsible + prepareAnsibleExecution + ansiblePrepResult
  - ansible-params.go  — mergeAnsibleParameters, parseStringParams,
                         convertStringValue, mergeParamMaps,
                         convertMapToAnsibleParams (parameter merging)
  - ansible-export.go  — ExecuteAnsibleWithExport,
                         ExecuteAnsibleEncryptAndCommit (export variants)

Also drops the duplicate `splitHostsForInventory` from inventory.go and
points its lone caller to the canonical `splitHosts`, which is moved
from ansible.go into inventory.go (where the issue suggests it belongs).

All 12 kept function bodies are byte-identical to the originals
(verified via per-function md5 across the new files). `dagger develop`
+ `go build ./...` + `go vet ./...` all pass.

Refs #143 (P3).
