+++ vm/ansible.go
@@ -3,13 +3,7 @@ package main
 import (
 	"context"
 	"dagger/vm/internal/dagger"
-	"encoding/json"
 	"fmt"
-	"path/filepath"
-	"strconv"
-	"strings"
-
-	"gopkg.in/yaml.v3"
 )
 
 func (m *Vm) ExecuteAnsible(
@@ -72,162 +66,6 @@ func (m *Vm) ExecuteAnsible(
 			SSHUser:        sshUser,
 			SSHPassword:    sshPassword,
 		})
-}
-
-// mergeAnsibleParameters merges parameters from YAML file and string
-// String parameters have higher priority and override file parameters
-func (m *Vm) mergeAnsibleParameters(ctx context.Context, file *dagger.File, strParams string) (string, error) {
-	// Start with empty parameters
-	fileParams := make(map[string]interface{})
-
-	// Load parameters from YAML file if provided
-	if file != nil {
-		content, err := file.Contents(ctx)
-		if err != nil {
-			return "", fmt.Errorf("failed to read parameters file: %w", err)
-		}
-
-		if err := yaml.Unmarshal([]byte(content), &fileParams); err != nil {
-			return "", fmt.Errorf("failed to parse YAML parameters: %w", err)
-		}
-	}
-
-	// If no string parameters, return YAML params as key=value pairs
-	if strParams == "" {
-		if len(fileParams) == 0 {
-			return "", nil
-		}
-		return convertMapToAnsibleParams(fileParams), nil
-	}
-
-	// Parse string parameters (format: "key1=value1,key2=value2")
-	strParamMap := parseStringParams(strParams)
-
-	// Merge maps with string params taking priority
-	mergedParams := mergeParamMaps(fileParams, strParamMap)
-
-	// Convert to Ansible parameters format (key=value pairs)
-	return convertMapToAnsibleParams(mergedParams), nil
-}
-
-// parseStringParams parses string parameters in format "key1=value1,key2=value2"
-func parseStringParams(params string) map[string]interface{} {
-	result := make(map[string]interface{})
-
-	if params == "" {
-		return result
-	}
-
-	// Split by comma to get key-value pairs
-	pairs := strings.Split(params, ",")
-	for _, pair := range pairs {
-		pair = strings.TrimSpace(pair)
-		if pair == "" {
-			continue
-		}
-
-		// Split by first equals sign to handle values with equals signs
-		kv := strings.SplitN(pair, "=", 2)
-		if len(kv) == 2 {
-			key := strings.TrimSpace(kv[0])
-			value := strings.TrimSpace(kv[1])
-
-			// Try to detect if value should be converted to int or bool
-			convertedValue := convertStringValue(value)
-			result[key] = convertedValue
-		}
-	}
-
-	return result
-}
-
-// convertStringValue attempts to convert string values to appropriate types
-func convertStringValue(s string) interface{} {
-	// Try boolean
-	sLower := strings.ToLower(s)
-	if sLower == "true" {
-		return true
-	}
-	if sLower == "false" {
-		return false
-	}
-
-	// Try integer
-	if intVal, err := strconv.Atoi(s); err == nil {
-		return intVal
-	}
-
-	// Try float
-	if floatVal, err := strconv.ParseFloat(s, 64); err == nil {
-		return floatVal
-	}
-
-	// Return as string
-	return s
-}
-
-// mergeParamMaps merges two parameter maps with overrideMap taking priority
-func mergeParamMaps(base, override map[string]interface{}) map[string]interface{} {
-	result := make(map[string]interface{})
-
-	// Copy base map
-	for k, v := range base {
-		result[k] = v
-	}
-
-	// Apply overrides
-	for k, v := range override {
-		result[k] = v
-	}
-
-	return result
-}
-
-// convertMapToAnsibleParams converts a map to Ansible parameters string format
-// Returns "key1=value1 key2=value2" format (space-separated)
-func convertMapToAnsibleParams(params map[string]interface{}) string {
-	if len(params) == 0 {
-		return ""
-	}
-
-	var pairs []string
-	for k, v := range params {
-		// Format the value appropriately
-		switch val := v.(type) {
-		case string:
-			pairs = append(pairs, fmt.Sprintf("%s=%s", k, val))
-		case bool:
-			pairs = append(pairs, fmt.Sprintf("%s=%v", k, val))
-		case int, int32, int64, float32, float64:
-			pairs = append(pairs, fmt.Sprintf("%s=%v", k, val))
-		default:
-			// For complex types, convert to JSON string
-			jsonBytes, err := json.Marshal(val)
-			if err != nil {
-				// Fallback to string representation
-				pairs = append(pairs, fmt.Sprintf("%s=%v", k, val))
-			} else {
-				// Escape single quotes in JSON for shell
-				jsonStr := strings.ReplaceAll(string(jsonBytes), "'", "'\"'\"'")
-				pairs = append(pairs, fmt.Sprintf("%s='%s'", k, jsonStr))
-			}
-		}
-	}
-
-	// Return space-separated pairs
-	return strings.Join(pairs, " ")
-}
-
-// splitHosts splits comma-separated hosts and trims whitespace
-func splitHosts(hosts string) []string {
-	var result []string
-	for _, host := range strings.Split(hosts, ",") {
-		host = strings.TrimSpace(host)
-		if host != "" {
-			result = append(result, host)
-		}
-	}
-	return result
 }
 
 // ansiblePrepResult holds the prepared inputs for Ansible execution.
@@ -304,208 +142,4 @@ func (m *Vm) prepareAnsibleExecution(
 		requirements: requirements,
 		parameters:   finalParameters,
 	}, nil
-}
-
-// ExecuteAnsibleWithExport runs Ansible playbooks and exports specified files from the container.
-// Same parameters as ExecuteAnsible plus exportPaths (comma-separated file paths to extract).
-func (m *Vm) ExecuteAnsibleWithExport(
-	ctx context.Context,
-	// +optional
-	src *dagger.Directory,
-	playbooks string,
-	// Comma-separated list of file paths to export from the Ansible container
-	exportPaths string,
-	// +optional
-	requirements *dagger.File,
-	// +optional
-	inventory *dagger.File,
-	// Comma-separated list of hosts (e.g., "192.168.1.10,192.168.1.11")
-	// Used to generate inventory if inventory file is not provided
-	// +optional
-	hosts string,
-	// +optional
-	parameters string,
-	// Path to a YAML file containing parameters (lower priority)
-	// +optional
-	parametersFile *dagger.File,
-	// +optional
-	vaultAppRoleID *dagger.Secret,
-	// +optional
-	vaultSecretID *dagger.Secret,
-	// +optional
-	vaultURL *dagger.Secret,
-	// +optional
-	sshUser *dagger.Secret,
-	// +optional
-	sshPassword *dagger.Secret,
-	// +optional
-	// +default="https://raw.githubusercontent.com/stuttgart-things/ansible/refs/heads/main/templates/requirements.yaml.tmpl"
-	requirementsTemplate string,
-	// +optional
-	// +default="https://raw.githubusercontent.com/stuttgart-things/ansible/refs/heads/main/templates/requirements-data.yaml"
-	requirementsData string,
-	// Inventory type: "simple" (default [all] group) or "cluster" (master/worker groups)
-	// +optional
-	// +default="simple"
-	inventoryType string,
-) (*dagger.Directory, error) {
-
-	prep, err := m.prepareAnsibleExecution(ctx, src, requirements, inventory, hosts, parameters, parametersFile, requirementsTemplate, requirementsData, inventoryType)
-	if err != nil {
-		return nil, err
-	}
-
-	// EXECUTE ANSIBLE AND EXPORT FILES
-	exportDir := dag.Ansible().ExecuteAndExport(
-		playbooks,
-		exportPaths,
-		dagger.AnsibleExecuteAndExportOpts{
-			Src:            prep.src,
-			Inventory:      prep.inventory,
-			Parameters:     prep.parameters,
-			VaultAppRoleID: vaultAppRoleID,
-			VaultSecretID:  vaultSecretID,
-			VaultURL:       vaultURL,
-			Requirements:   prep.requirements,
-			SSHUser:        sshUser,
-			SSHPassword:    sshPassword,
-		})
-
-	return exportDir, nil
-}
-
-// ExecuteAnsibleEncryptAndCommit runs Ansible playbooks, extracts files from the container,
-// encrypts them with SOPS, and commits the encrypted files to a Git repository.
-func (m *Vm) ExecuteAnsibleEncryptAndCommit(
-	ctx context.Context,
-	// +optional
-	src *dagger.Directory,
-	playbooks string,
-	// Comma-separated list of file paths to export from the Ansible container
-	exportPaths string,
-	// +optional
-	requirements *dagger.File,
-	// +optional
-	inventory *dagger.File,
-	// Comma-separated list of hosts (e.g., "192.168.1.10,192.168.1.11")
-	// +optional
-	hosts string,
-	// +optional
-	parameters string,
-	// +optional
-	parametersFile *dagger.File,
-	// +optional
-	vaultAppRoleID *dagger.Secret,
-	// +optional
-	vaultSecretID *dagger.Secret,
-	// +optional
-	vaultURL *dagger.Secret,
-	// +optional
-	sshUser *dagger.Secret,
-	// +optional
-	sshPassword *dagger.Secret,
-	// +optional
-	// +default="https://raw.githubusercontent.com/stuttgart-things/ansible/refs/heads/main/templates/requirements.yaml.tmpl"
-	requirementsTemplate string,
-	// +optional
-	// +default="https://raw.githubusercontent.com/stuttgart-things/ansible/refs/heads/main/templates/requirements-data.yaml"
-	requirementsData string,
-	// Inventory type: "simple" (default [all] group) or "cluster" (master/worker groups)
-	// +optional
-	// +default="simple"
-	inventoryType string,
-	// AGE public key for SOPS encryption
-	agePublicKey *dagger.Secret,
-	// File extension for SOPS encryption (e.g., "yaml", "json")
-	// +optional
-	// +default="yaml"
-	sopsFileExtension string,
-	// SOPS config file (.sops.yaml)
-	// +optional
-	sopsConfig *dagger.File,
-	// Git repository in "owner/repo" format
-	gitRepository string,
-	// Git branch name
-	// +optional
-	// +default="main"
-	gitBranch string,
-	// Git commit message
-	// +optional
-	// +default="Add encrypted files from Ansible execution"
-	gitCommitMessage string,
-	// Destination path within the git repository
-	// +optional
-	// +default="/"
-	gitDestinationPath string,
-	// GitHub token for authentication
-	gitToken *dagger.Secret,
-	// If non-empty, create this branch (from gitBranch as base) and commit there instead
-	// +optional
-	gitCreateBranch string,
-	// If true (and gitCreateBranch set), open a PR from the new branch back to gitBranch
-	// +optional
-	gitCreatePr bool,
-	// PR title (defaults to gitCommitMessage if empty)
-	// +optional
-	gitPrTitle string,
-	// Comma-separated list of target filenames for exported files (maps 1:1 to exportPaths)
-	// If not set, original filenames are used
-	// +optional
-	exportTargetNames string,
-) (string, error) {
-
-	// PHASE 1: Execute Ansible and export files
-	exportDir, err := m.ExecuteAnsibleWithExport(
-		ctx, src, playbooks, exportPaths, requirements, inventory, hosts, parameters, parametersFile,
-		vaultAppRoleID, vaultSecretID, vaultURL, sshUser, sshPassword,
-		requirementsTemplate, requirementsData, inventoryType,
-	)
-	if err != nil {
-		return "", fmt.Errorf("ansible execution failed: %w", err)
-	}
-
-	// PHASE 2: Encrypt each exported file with SOPS
-	entries, err := exportDir.Entries(ctx)
-	if err != nil {
-		return "", fmt.Errorf("failed to list exported files: %w", err)
-	}
-
-	// Build rename map from exportTargetNames if provided
-	renameMap := make(map[string]string)
-	if exportTargetNames != "" {
-		targetNames := strings.Split(exportTargetNames, ",")
-		exportPathsList := strings.Split(exportPaths, ",")
-		if len(targetNames) != len(exportPathsList) {
-			return "", fmt.Errorf("exportTargetNames count (%d) must match exportPaths count (%d)", len(targetNames), len(exportPathsList))
-		}
-		for i, ep := range exportPathsList {
-			baseName := filepath.Base(strings.TrimSpace(ep))
-			renameMap[baseName] = strings.TrimSpace(targetNames[i])
-		}
-	}
-
-	encryptedDir := dag.Directory()
-	for _, entry := range entries {
-		plaintextFile := exportDir.File(entry)
-
-		encryptedContent, err := m.EncryptFile(ctx, agePublicKey, plaintextFile, sopsFileExtension, sopsConfig)
-		if err != nil {
-			return "", fmt.Errorf("failed to encrypt file %s: %w", entry, err)
-		}
-
-		targetName := entry
-		if mapped, ok := renameMap[entry]; ok {
-			targetName = mapped
-		}
-
-		encryptedDir = encryptedDir.WithNewFile(targetName, encryptedContent)
-	}
-
-	// PHASE 3: Commit encrypted files to Git
-	result, err := m.CommitToGit(ctx, encryptedDir, gitRepository, gitBranch, gitCommitMessage, gitDestinationPath, gitToken, gitCreateBranch, gitCreatePr, gitPrTitle)
-	if err != nil {
-		return "", fmt.Errorf("git commit failed: %w", err)
-	}
-
-	return result, nil
 }
